Move view access tests to capabilities
