+++ jetro-core/src/exec/pipeline/capability.rs
@@ -165,7 +165,8 @@ pub(crate) enum SourceAccessMode {
 
 #[cfg(test)]
 mod source_capability_tests {
-    use super::{SourceAccessMode, SourceCapabilities};
+    use super::{SourceAccessMode, SourceCapabilities, ViewStageCapability};
+    use crate::data::value::Val;
     use crate::plan::demand::{FieldDemand, FieldSet, PullDemand};
     use std::sync::Arc;
 
@@ -212,6 +213,41 @@ mod source_capability_tests {
             caps.choose_access(PullDemand::LastInput(1)),
             SourceAccessMode::IndexedFromEnd(0)
         );
+    }
+
+    #[test]
+    fn selective_view_prefix_demotes_indexed_last_to_reverse_scan() {
+        let access = SourceCapabilities::VIEW_ARRAY.choose_view_access(
+            PullDemand::LastInput(1),
+            &[ViewStageCapability::Filter { kernel: 0 }],
+        );
+
+        assert_eq!(access, SourceAccessMode::Reverse { outputs: 1 });
+    }
+
+    #[test]
+    fn selective_view_prefix_demotes_indexed_last_to_forward_without_reverse() {
+        let caps = SourceCapabilities {
+            reverse_stream: false,
+            ..SourceCapabilities::VIEW_ARRAY
+        };
+
+        let access = caps.choose_view_access(
+            PullDemand::LastInput(1),
+            &[ViewStageCapability::RemoveValue(Val::Int(2))],
+        );
+
+        assert_eq!(access, SourceAccessMode::Forward);
+    }
+
+    #[test]
+    fn cardinality_preserving_view_prefix_keeps_indexed_last_seek() {
+        let access = SourceCapabilities::VIEW_ARRAY.choose_view_access(
+            PullDemand::LastInput(1),
+            &[ViewStageCapability::Map { kernel: 0 }],
+        );
+
+        assert_eq!(access, SourceAccessMode::IndexedFromEnd(0));
     }
 
     #[test]
